Toevoegen backend permissies campagnes

+++ app/Http/Controllers/Api/Campaign/CampaignController.php
@@ -143,6 +143,7 @@ class CampaignController extends ApiController
 
     public function associateOpportunity(Campaign $campaign, Opportunity $opportunity)
     {
+        $this->authorize('manage', Campaign::class);
         $opportunity->campaign()->associate($campaign);
         $opportunity->save();
 
@@ -151,6 +152,7 @@ class CampaignController extends ApiController
 
     public function dissociateOpportunity(Opportunity $opportunity)
     {
+        $this->authorize('manage', Campaign::class);
         $opportunity->campaign()->dissociate();
         $opportunity->save();
 
@@ -159,6 +161,7 @@ class CampaignController extends ApiController
 
     public function attachResponse(Campaign $campaign, Contact $contact)
     {
+        $this->authorize('manage', Campaign::class);
         $campaignResponse = new CampaignResponse([
             'campaign_id' => $campaign->id,
             'contact_id' => $contact->id,
@@ -169,17 +172,20 @@ class CampaignController extends ApiController
 
     public function detachResponse(Campaign $campaign, Contact $contact)
     {
+        $this->authorize('manage', Campaign::class);
         $campaign->responses()->where('contact_id', $contact->id)->delete();
         $campaign->save();
     }
 
     public function attachOrganisation(Campaign $campaign, Organisation $organisation)
     {
+        $this->authorize('manage', Campaign::class);
         $campaign->organisations()->attach($organisation);
     }
 
     public function detachOrganisation(Campaign $campaign, Organisation $organisation)
     {
+        $this->authorize('manage', Campaign::class);
         $campaign->organisations()->detach($organisation);
     }
 
@@ -190,6 +196,7 @@ class CampaignController extends ApiController
 
     public function associateOwner(Campaign $campaign, User $user)
     {
+        $this->authorize('manage', Campaign::class);
         $campaign->ownedBy()->associate($user);
         $campaign->save();
     }
